sound: remove floating point operations and use NUKED OPL
* NUKED OPL does not contain any floating point operation

+++ include/players.h
@@ -34,7 +34,7 @@ typedef struct tagAUDIOPLAYER
     INT                        iMusic;  \
     BOOL                       fLoop; \
 	VOID (*Shutdown)(VOID*); \
-	BOOL (*Play)(VOID*, INT, BOOL, float); \
+	BOOL (*Play)(VOID*, INT, BOOL, INT); \
 	VOID (*FillBuffer)(VOID*, LPBYTE, INT)
 
 	AUDIOPLAYER_COMMONS;
